Improve code. Fix UART wrong parity

+++ utils_misc/include/utils_misc/int_macros.h
@@ -2,6 +2,7 @@
  * \file utils_misc/int_macros.h
  * \brief Basic helper macros for integers
  */
+#include <stdint.h>
 
 #pragma once
 
@@ -14,10 +15,10 @@
 
 #define IS_IN_RANGE(lowLim, val, highLim) (((lowLim) <= (val)) && ((val) <= (highLim)))
 
-#define GET_BYTE_0(src) (((src)>>0)&0xff)
-#define GET_BYTE_1(src) (((src)>>8)&0xff)
-#define GET_BYTE_2(src) (((src)>>16)&0xff)
-#define GET_BYTE_3(src) (((src)>>24)&0xff)
+#define GET_BYTE_0(src) (uint8_t)(((src)>>0)&0xff)
+#define GET_BYTE_1(src) (uint8_t)(((src)>>8)&0xff)
+#define GET_BYTE_2(src) (uint8_t)(((src)>>16)&0xff)
+#define GET_BYTE_3(src) (uint8_t)(((src)>>24)&0xff)
 
 #define PUT_LOW_NIBBLE(dst, val) ((dst) = ((dst)&0xf0)|((val)&0x0f))
 #define PUT_HIGH_NIBBLE(dst, val) ((dst) = (((val)<<4)&0xf0)|((dst)&0x0f))
